refactor: rename hook file names
