+++ apps/calendar/src/components/dayGridMonth/dayGridMonth.tsx
@@ -18,8 +18,8 @@ import { cls, toPercent } from '@src/helpers/css';
 import { getVisibleEventCollection } from '@src/helpers/events';
 import { createGridPositionFinder, getRenderedEventUIModels } from '@src/helpers/grid';
 import { dayGridMonthSelectionHelper } from '@src/helpers/gridSelection';
-import { useDOMNode } from '@src/hooks/common/domNode';
-import { useGridSelection } from '@src/hooks/gridSelection/gridSelection';
+import { useDOMNode } from '@src/hooks/common/useDOMNode';
+import { useGridSelection } from '@src/hooks/gridSelection/useGridSelection';
 import { calendarSelector } from '@src/selectors';
 import TZDate from '@src/time/date';
 import { getSize } from '@src/utils/dom';
