+++ apps/calendar/src/components/events/horizontalEvent.tsx
@@ -8,8 +8,8 @@ import { useEventBus } from '@src/contexts/eventBus';
 import { useLayoutContainer } from '@src/contexts/layoutContainer';
 import { cls, toPercent, toPx } from '@src/helpers/css';
 import { DRAGGING_TYPE_CREATORS } from '@src/helpers/drag';
-import { useDrag } from '@src/hooks/common/drag';
-import { useTransientUpdate } from '@src/hooks/common/transientUpdate';
+import { useDrag } from '@src/hooks/common/useDrag';
+import { useTransientUpdate } from '@src/hooks/common/useTransientUpdate';
 import EventUIModel from '@src/model/eventUIModel';
 import { dndSelector, optionsSelector } from '@src/selectors';
 import { DraggingState } from '@src/slices/dnd';
