+++ apps/calendar/src/components/events/timeEvent.tsx
@@ -6,8 +6,8 @@ import { useDispatch } from '@src/contexts/calendarStore';
 import { useLayoutContainer } from '@src/contexts/layoutContainer';
 import { cls, toPercent } from '@src/helpers/css';
 import { DRAGGING_TYPE_CREATORS } from '@src/helpers/drag';
-import { useDrag } from '@src/hooks/common/drag';
-import { useTransientUpdate } from '@src/hooks/common/transientUpdate';
+import { useDrag } from '@src/hooks/common/useDrag';
+import { useTransientUpdate } from '@src/hooks/common/useTransientUpdate';
 import EventUIModel from '@src/model/eventUIModel';
 import { dndSelector } from '@src/selectors';
 import { DraggingState } from '@src/slices/dnd';
