+++ apps/calendar/src/components/panelResizer.tsx
@@ -4,7 +4,7 @@ import { useRef, useState } from 'preact/hooks';
 import { useDispatch } from '@src/contexts/calendarStore';
 import { cls } from '@src/helpers/css';
 import { DRAGGING_TYPE_CONSTANTS } from '@src/helpers/drag';
-import { useDrag } from '@src/hooks/common/drag';
+import { useDrag } from '@src/hooks/common/useDrag';
 
 import { StyleProp } from '@t/components/common';
 import { AlldayEventCategory } from '@t/panel';
