+++ apps/calendar/src/components/popup/eventStateSelector.tsx
@@ -3,7 +3,7 @@ import { h } from 'preact';
 import { PopupSection } from '@src/components/popup/popupSection';
 import { StateDropdownMenu } from '@src/components/popup/stateDropdownMenu';
 import { cls } from '@src/helpers/css';
-import { useDropdownState } from '@src/hooks/common/dropdownState';
+import { useDropdownState } from '@src/hooks/common/useDropdownState';
 import { FormStateActionType, FormStateDispatcher } from '@src/hooks/popup/formState';
 
 import { EventState } from '@t/events';
