+++ apps/calendar/src/hooks/dayGridMonth/useDayGridMonthEventMove.ts
@@ -1,8 +1,8 @@
 import { useEffect, useMemo } from 'preact/hooks';
 
 import { useDispatch, useStore } from '@src/contexts/calendarStore';
-import { useCurrentPointerPositionInGrid } from '@src/hooks/event/currentPointerPositionInGrid';
-import { useDraggingEvent } from '@src/hooks/event/draggingEvent';
+import { useCurrentPointerPositionInGrid } from '@src/hooks/event/useCurrentPointerPositionInGrid';
+import { useDraggingEvent } from '@src/hooks/event/useDraggingEvent';
 import { isNotDraggingSelector } from '@src/selectors/dnd';
 import TZDate from '@src/time/date';
 import { getDateDifference, MS_PER_DAY } from '@src/time/datetime';
