+++ apps/calendar/src/hooks/dayGridMonth/useDayGridMonthEventResize.ts
@@ -3,9 +3,9 @@ import { useCallback, useEffect, useMemo, useState } from 'preact/hooks';
 import { KEY } from '@src/constants/keyboard';
 import { useDispatch, useStore } from '@src/contexts/calendarStore';
 import { getGridDateIndex, getRenderedEventUIModels } from '@src/helpers/grid';
-import { useKeydownEvent } from '@src/hooks/common/keydownEvent';
-import { useCurrentPointerPositionInGrid } from '@src/hooks/event/currentPointerPositionInGrid';
-import { useDraggingEvent } from '@src/hooks/event/draggingEvent';
+import { useKeydownEvent } from '@src/hooks/common/useKeydownEvent';
+import { useCurrentPointerPositionInGrid } from '@src/hooks/event/useCurrentPointerPositionInGrid';
+import { useDraggingEvent } from '@src/hooks/event/useDraggingEvent';
 import EventUIModel from '@src/model/eventUIModel';
 import { isNotDraggingSelector } from '@src/selectors/dnd';
 import TZDate from '@src/time/date';
