+++ apps/calendar/src/hooks/event/useCurrentPointerPositionInGrid.ts
@@ -1,6 +1,6 @@
 import { useCallback, useState } from 'preact/hooks';
 
-import { useTransientUpdate } from '@src/hooks/common/transientUpdate';
+import { useTransientUpdate } from '@src/hooks/common/useTransientUpdate';
 import { dndSelector } from '@src/selectors';
 import { isPresent } from '@src/utils/type';
