+++ apps/calendar/src/hooks/gridSelection/useGridSelection.ts
@@ -3,8 +3,8 @@ import { useCallback, useEffect, useRef, useState } from 'preact/hooks';
 import { useDispatch, useStore } from '@src/contexts/calendarStore';
 import { useEventBus } from '@src/contexts/eventBus';
 import { DRAGGING_TYPE_CREATORS } from '@src/helpers/drag';
-import { useDrag } from '@src/hooks/common/drag';
-import { useTransientUpdate } from '@src/hooks/common/transientUpdate';
+import { useDrag } from '@src/hooks/common/useDrag';
+import { useTransientUpdate } from '@src/hooks/common/useTransientUpdate';
 import { dndSelector } from '@src/selectors';
 import { DraggingState } from '@src/slices/dnd';
 import { GridSelectionType } from '@src/slices/gridSelection';
