+++ apps/calendar/src/hooks/timeGrid/timeGridEventMove.spec.tsx
@@ -5,7 +5,7 @@ import userEvent from '@testing-library/user-event';
 import { initCalendarStore, StoreProvider } from '@src/contexts/calendarStore';
 import { DRAGGING_TYPE_CREATORS } from '@src/helpers/drag';
 import { createGridPositionFinder, createTimeGridData, getWeekDates } from '@src/helpers/grid';
-import { useDrag } from '@src/hooks/common/drag';
+import { useDrag } from '@src/hooks/common/useDrag';
 import { useTimeGridEventMove } from '@src/hooks/timeGrid/timeGridEventMove';
 import EventModel from '@src/model/eventModel';
 import EventUIModel from '@src/model/eventUIModel';
